[lldb][Breakpoint] Prevent crash when resolving regex breakpoint on functions with asm declaration
Currently, a function whose mangled name got
changed using an `asm()` statement on its
declaration will cause lldb to crash when
resolving restricted regex breakpoints.

The crash occurs because we previously assumed
that all functions in the SymbolContext will
cleanly demangle into a function name that we
can compare against the restrictions list.
However, with the `asm()` attribute applied
the mangled name may not conform to the
supported mangling schemes and we return
a nullptr.

The fix is to use the fallback parameter to
`ConstString::AsCString()`.

**Testing**

* Added API test

Differential Revision: https://reviews.llvm.org/D131693