[lldb][test] Add check for Xcode binutils version to test-runner
Some tests in LLDB require the use of binutils tools. On Darwin,
we need to use the tools that ship with Xcode. Otherwise we risk
crashing in unpredictable ways.

This patch adds a check for this and bails early if the user has
a non-Xcode binutils in their path before the Xcode ones.