Fix Objective-C tests: Make sure the ObjCInterfaceDecl::hasDefinition checks for out-dated decls
A previous patch made sure we set the Data.setInt(0) not only when
modules are enabled but also when we create Decls in the debugger.
This aligns with the overall strategy of the TC rework to move
completion to how the modules infrastructure does it.

This patch makes sure we set the DebuggerSupport on all possible
ASTContexts we create Objective-C decls on (we already set it in
ClangExpressionParser) such that when we ask an objective-c decl for
its definition (via `getDefinition`), we actually pull the definition
in via the lazy redecl infrastructure.