[clang][ASTImporter] Add a 'Message' member to ASTImportError use it throught ASTImporter
This patch adds a new `ASTImportError::Message` which is used
to more accurately describe a failure when `ASTImportError::log`
is called. We then set this error message throughout ASTImporter.

The motivation for this is that in upcoming patches LLDB will
check `ASTImportError`s and log them on failures. At the moment
these logs wouldn't be helpful since we wouldn't be able to
differentiate between the different conflicts that occur
during a single import process.