[lldb][test] Skip *-dbg-info-content API tests
These tests started failing on the public build-bots recently
with following error:
```
AssertionError: 'error: Couldn't lookup symbols:
  __ZNSt3__122__libcpp_verbose_abortEPKcz
' is not success
```

We've seen this previously when the SDKs we used to compile the
`std` module differ from the test program. (see D146714, rdar://107052293)

Skip these tests on older MacOS versions for now.

This is possibly related to the recent `std` module changes in D144322.