[lldb][XcodeSDK] Simplify logic that adjusts sysroot during XcodeSDK merging
The `DW_AT_APPLE_sdk` should always be equal to the filename of the `DW_AT_LLVM_sysroot`. We can use this property to simplify `XcodeSDK::Merge` to no longer manually adjust the sysroot filename. Instead we simply update the sysroot filename with merged SDK name.

This should be an NFC change.