[lldb][Expression] Remove redundant DeclVendor modules lookup
`ClangExpressionDeclMap::FindExternalVisibleDecls` will perform a
`LookupInModulesDeclVendor` anyway if `LookupFunction` didn't find any
functions in debug-info. So there is no need to duplicate this code
inside `LookupFunctionp`. The order of how we lookup functions won't
change. We still first perform a `FindFunctions` and if that doesn't
work we ask the `DeclVendor`.

The non-NFC part of this patch is that `LookupInModulesDeclVendor` will
only look for 1 match. While `LookupFunction` would look at all the
functions we found in the modules. I couldn't find a compelling reason
for why we would only look at 1 decl, so I now loop over all decls we
found.