[lldb][Expression] Register all functions find in Modules
This is what the module lookup in `LookupFunction` already does. This
code-path is (IIUC) dead for function lookups anyway. I'm trying to
consolidate the lookup in `LookupFunction` and
`LookupInModulesDeclVendor`, and this patch gets us slightly closer.

I tried adding a test (e.g., where we don't have debug-info for a
function but it is available in a Clang module). However, those are
handled by `LookupFunction` already. So there's no good way of adding
coverage here, which shouldn't matter because this is unused (for
function lookups).